// src/types/generated/CanonicalRowSet.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CanonicalBuildRow } from "./CanonicalBuildRow";
import type { CanonicalMatchupRow } from "./CanonicalMatchupRow";
import type { CanonicalRateRow } from "./CanonicalRateRow";

export type CanonicalRowSet = { region: string, rates: Array<CanonicalRateRow>, matchups: Array<CanonicalMatchupRow>, builds: Array<CanonicalBuildRow>, };